+++ slangpy/native_refl/__init__.py
@@ -1,3 +1,1 @@
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-
-"""Transitional native semantic reflection bindings for SlangPy internals."""
