+++ slangpy/tests/slangpy_tests/test_reflection2.py
@@ -2,10 +2,12 @@
 
 import pytest
 
-from slangpy import DeviceType, TypeReflection
+from slangpy import DeviceType, Tensor, TextureUsage, TypeReflection
 import slangpy.reflection as r
-from slangpy.reflection.reflectiontypes import is_float
+from slangpy.core.enums import IOType
 from slangpy.core.function import Function
+from slangpy.reflection.lookup import resolve_element_type, resolve_program_layout
+from slangpy.reflection.reflectiontypes import is_float
 from slangpy.testing import helpers
 
 from typing import Any, Callable
@@ -20,11 +22,18 @@ float foo_ol(float a, float b) { return a+b; }
 float foo_generic<T>(T a) { return 0; }
 struct Foo
 {
+    float3 value;
     int bar(float a) {}
 }
 
 struct GenericType<A, int N> {}
 struct BoolGenericType<let Enabled: bool> {}
+
+void update(inout float value, out float result, no_diff in float weight) {
+    result = value * weight;
+}
+
+void use_textures(Texture2D<float4> texture, RWTexture2D<float4> rw_texture) {}
 
 """
 
@@ -93,6 +102,166 @@ def test_method(device_type: DeviceType):
     assert res.parameters[0].type == layout.scalar_type(TypeReflection.ScalarType.float32)
     assert res.have_return_value
     assert res.return_type == layout.scalar_type(TypeReflection.ScalarType.int32)
+
+
+@pytest.mark.parametrize("device_type", helpers.DEFAULT_DEVICE_TYPES)
+def test_module_and_struct_reflection_storage(device_type: DeviceType):
+    device = helpers.get_device(device_type)
+    module = helpers.create_module(device, MODULE)
+
+    assert not hasattr(module, "module")
+    assert "layout" not in module.__dict__
+    assert "device_module" not in module.__dict__
+    assert module.layout.find_type_by_name("Foo") is not None
+
+    struct = module.Foo.as_struct()
+
+    assert struct.module is module
+    assert struct.struct is struct.type
+    assert "module" not in struct.__dict__
+    assert "struct" not in struct.__dict__
+    assert struct.struct.type_reflection.full_name == "Foo"
+    assert struct.type_reflection.full_name == "Foo"
+    assert not hasattr(struct.struct, "reflection")
+    assert not hasattr(struct, "reflection")
+
+    tensor = Tensor.empty(device, (2,), dtype=struct)
+    assert tensor.dtype is struct.struct
+
+
+@pytest.mark.parametrize("device_type", helpers.DEFAULT_DEVICE_TYPES)
+def test_reflection_layout_tracks_hot_reload_generation(device_type: DeviceType):
+    device = helpers.get_device(device_type)
+    module = helpers.create_module(device, MODULE)
+    layout = r.SlangProgramLayout(module.device_module.layout)
+
+    generation = layout.generation
+    layout.on_hot_reload(module.device_module.layout)
+    module.on_hot_reload()
+
+    assert layout.generation == generation + 1
+    assert module.layout.find_type_by_name("Foo") is not None
+
+
+@pytest.mark.parametrize("device_type", helpers.DEFAULT_DEVICE_TYPES)
+def test_reflection_layout_creates_semantic_types(device_type: DeviceType):
+    device = helpers.get_device(device_type)
+    module = helpers.create_module(device, MODULE)
+    layout = r.SlangProgramLayout(module.device_module.layout)
+
+    float_type = layout.require_type_by_name("float")
+    assert isinstance(float_type, r.ScalarType)
+    assert float_type.full_name == "float"
+    assert float_type.shape == ()
+
+    void_type = layout.scalar_type(TypeReflection.ScalarType.void)
+    assert isinstance(void_type, r.VoidType)
+    assert void_type.full_name == "void"
+
+    vector_type = layout.require_type_by_name("vector<float,3>")
+    assert isinstance(vector_type, r.VectorType)
+    assert vector_type.element_type is float_type
+    assert vector_type.num_elements == 3
+    assert vector_type.shape == (3,)
+
+    matrix_type = layout.require_type_by_name("matrix<float,3,2>")
+    assert isinstance(matrix_type, r.MatrixType)
+    assert matrix_type.inner_element_type is float_type
+    assert matrix_type.shape == (3, 2)
+
+    array_type = layout.require_type_by_name("float[4]")
+    assert isinstance(array_type, r.ArrayType)
+    assert array_type.element_type is float_type
+    assert array_type.array_shape == (4,)
+
+    struct_type = layout.require_type_by_name("Foo")
+    assert isinstance(struct_type, r.StructType)
+    assert struct_type.full_name == "Foo"
+    assert isinstance(struct_type.fields["value"], r.SlangField)
+    assert struct_type.fields["value"].type is vector_type
+
+    tensor_type = layout.tensor_type(
+        float_type,
+        2,
+        access=r.TensorAccess.read_write,
+        tensor_kind=r.TensorType.tensor,
+    )
+    assert isinstance(tensor_type, r.ITensorType)
+    assert tensor_type.dtype is float_type
+    assert tensor_type.dims == 2
+    assert tensor_type.readable
+    assert tensor_type.writable
+    assert tensor_type.shape == (-1, -1)
+
+    texture_function = layout.require_function_by_name("use_textures")
+    texture_type = texture_function.parameters[0].type
+    rw_texture_type = texture_function.parameters[1].type
+    assert isinstance(texture_type, r.TextureType)
+    assert texture_type.usage == TextureUsage.shader_resource
+    assert isinstance(rw_texture_type, r.TextureType)
+    assert rw_texture_type.usage == TextureUsage.unordered_access
+
+
+@pytest.mark.parametrize("device_type", helpers.DEFAULT_DEVICE_TYPES)
+def test_reflection_layout_creates_function_metadata(device_type: DeviceType):
+    device = helpers.get_device(device_type)
+    module = helpers.create_module(device, MODULE)
+    layout = r.SlangProgramLayout(module.device_module.layout)
+
+    float_type = layout.require_type_by_name("float")
+    struct_type = layout.require_type_by_name("Foo")
+
+    function = layout.require_function_by_name("foo2")
+    assert isinstance(function, r.SlangFunction)
+    assert function.name == "foo2"
+    assert function.return_type is float_type
+    assert function.have_return_value
+
+    parameters = function.parameters
+    assert len(parameters) == 2
+    assert isinstance(parameters[0], r.SlangParameter)
+    assert parameters[0].name == "a"
+    assert parameters[0].index == 0
+    assert parameters[0].type is float_type
+    assert parameters[0].io_type == IOType.inn
+
+    update = layout.require_function_by_name("update")
+    assert update.parameters[0].io_type == IOType.inout
+    assert update.parameters[1].io_type == IOType.out
+    assert update.parameters[2].no_diff
+
+    method = layout.require_function_by_name_in_type(struct_type, "bar")
+    assert method.this_type is struct_type
+    assert method.full_name == "bar"
+
+
+@pytest.mark.parametrize("device_type", helpers.DEFAULT_DEVICE_TYPES)
+def test_reflection_lookup_resolves_builtin_and_struct_types(device_type: DeviceType):
+    device = helpers.get_device(device_type)
+
+    builtin_layout = resolve_program_layout(device, None, None)
+    assert builtin_layout.find_type_by_name("float") is not None
+
+    float_type = resolve_element_type(builtin_layout, "float")
+    assert isinstance(float_type, r.ScalarType)
+    assert float_type.full_name == "float"
+
+    generation = builtin_layout.generation
+    device.reload_all_programs()
+    assert builtin_layout.generation == generation + 1
+
+    module = helpers.create_module(device, MODULE)
+    struct = module.Foo.as_struct()
+    struct_layout = resolve_program_layout(device, struct, None)
+    assert struct_layout.find_type_by_name("Foo") is not None
+
+    python_struct_type = module.layout.require_type_by_name("Foo")
+    assert resolve_element_type(struct_layout, struct).full_name == "Foo"
+    assert resolve_element_type(struct_layout, python_struct_type).full_name == "Foo"
+    assert (
+        resolve_element_type(struct_layout, python_struct_type.buffer_layout.reflection).full_name
+        == "Foo"
+    )
 
 
 @pytest.mark.parametrize("device_type", helpers.DEFAULT_DEVICE_TYPES)
